requirements.txt
alembic==1.6.2
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-Migrate==3.0.0
Flask-Script==2.0.6
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
simplejson==3.17.2
six==1.16.0
SQLAlchemy==1.4.15
watchdog==2.1.1
Werkzeug==2.0.0
WTForms==2.3.3
